# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "abacus",
    "description": "",
    "license": {
      "name": "EUPL-1.2",
      "identifier": "EUPL-1.2"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/polling_stations/{polling_station_id}/investigation": {
      "put": {
        "summary": "Update an investigation for a polling station",
        "operationId": "polling_station_investigation_update",
        "parameters": [
          {
            "name": "polling_station_id",
            "in": "path",
            "description": "Polling station database id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollingStationInvestigationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Polling station investigation updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollingStationInvestigation"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Investigation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Request cannot be completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an investigation for a polling station",
        "operationId": "polling_station_investigation_create",
        "parameters": [
          {
            "name": "polling_station_id",
            "in": "path",
            "description": "Polling station database id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollingStationInvestigationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Polling station investigation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollingStationInvestigation"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Polling station not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Request cannot be completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "my_auth": [
              "read:items",
              "edit:items"
            ]
          },
          {
            "token_jwt": []
          }
        ]
      },
      "delete": {
        "summary": "Delete an investigation for a polling station",
        "operationId": "polling_station_investigation_delete",
        "parameters": [
          {
            "name": "polling_station_id",
            "in": "path",
            "description": "Polling station database id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Polling station investigation deleted successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Investigation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/polling_stations/{polling_station_id}/investigation/conclude": {
      "post": {
        "summary": "Conclude an investigation for a polling station",
        "operationId": "polling_station_investigation_conclude",
        "parameters": [
          {
            "name": "polling_station_id",
            "in": "path",
            "description": "Polling station database id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollingStationInvestigationConcludeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Polling station investigation concluded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollingStationInvestigation"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Investigation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Request cannot be completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/polling_stations/{polling_station_id}/investigation/download_corrigendum_pdf": {
      "get": {
        "summary": "Download a corrigendum for a polling station",
        "operationId": "polling_station_investigation_download_corrigendum_pdf",
        "parameters": [
          {
            "name": "polling_station_id",
            "in": "path",
            "description": "Polling station database id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "attachment; filename=\"filename.pdf\""
              }
            },
            "content": {
              "application/pdf": {}
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Request cannot be completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorReference": {
        "type": "string",
        "description": "Error reference used to show the corresponding error message to the end-user",
        "enum": [
          "AirgapViolation",
          "AlreadyInitialised",
          "CommitteeSessionPaused",
          "DatabaseError",
          "DataEntryAlreadyClaimed",
          "DataEntryAlreadyFinalised",
          "DataEntryCannotBeDeleted",
          "DataEntryNotAllowed",
          "EmlImportError",
          "EntryNotFound",
          "EntryNotUnique",
          "Forbidden",
          "InternalServerError",
          "InvalidCommitteeSessionStatus",
          "InvalidData",
          "InvalidHash",
          "InvalidJson",
          "InvalidPassword",
          "InvalidPoliticalGroup",
          "InvalidSession",
          "InvalidStateTransition",
          "InvalidUsernameOrPassword",
          "InvalidVoteCandidate",
          "InvalidVoteGroup",
          "InvalidXml",
          "InvestigationHasDataEntryOrResult",
          "InvestigationRequiresCorrectedResults",
          "NotInitialised",
          "OwnAccountCannotBeDeleted",
          "PasswordRejection",
          "PdfGenerationError",
          "PollingStationRepeated",
          "PollingStationValidationErrors",
          "RequestPayloadTooLarge",
          "Unauthorized",
          "UsernameNotUnique",
          "UserNotFound"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Response structure for errors",
        "required": [
          "error",
          "fatal",
          "reference"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "fatal": {
            "type": "boolean"
          },
          "reference": {
            "$ref": "#/components/schemas/ErrorReference"
          }
        },
        "additionalProperties": false
      },
...
      "PollingStationInvestigation": {
        "type": "object",
        "required": [
          "polling_station_id",
          "reason"
        ],
        "properties": {
          "corrected_results": {
            "type": "boolean"
          },
          "findings": {
            "type": "string"
          },
          "polling_station_id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "reason": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PollingStationInvestigationConcludeRequest": {
        "type": "object",
        "required": [
          "findings",
          "corrected_results"
        ],
        "properties": {
          "corrected_results": {
            "type": "boolean"
          },
          "findings": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PollingStationInvestigationCreateRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PollingStationInvestigationUpdateRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "accept_data_entry_deletion": {
            "type": "boolean"
          },
          "corrected_results": {
            "type": "boolean"
          },
          "findings": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "my_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ABACUS_SESSION"
      }
    }
  }
}